+++ src/main/java/com/g93/be/service/MailTestService.java
@@ -1,0 +1,8 @@
+package com.g93.be.service;
+
+import com.g93.be.dto.SmtpTestResponse;
+
+public interface MailTestService {
+
+    SmtpTestResponse queueTestEmail(String recipient);
+}